added patient medical history

+++ src/main/java/com/curahealthyme/repo/DiseaseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.curahealthyme.model.Disease;
 
-public interface DiseaseRepository  extends JpaRepository<Disease, Integer>{
+public interface DiseaseRepository  extends JpaRepository<Disease, Long>{
 
 }
